hybrid_framework/requirements.txt
selenium
pytest
behave
requests
appium-python-client
transformers